Created a dataframe for asian restaurants.
Wrote some code to list all the genres of food in the restaurants dataframe. Then I used the categories of all asian restaurants and put it in a dataframe. Now I should have a dataframe of ALL asian restaurants with their review count, stars, etc. Now we just need to do the same thing for every other genre.